Split up /lib
I split up fields to separate files. Further notes at the source.

Note that it seems to be possible that `this.props.type` is not defined
for Selection always (ie. for enums). I made it to default to string in
this case. Perhaps there is a nicer way to fix this at a higher level.